refactor field
