+++ be/src/olap/push_handler.cpp
@@ -24,7 +24,6 @@
 
 #include "common/object_pool.h"
 #include "common/status.h"
-#include "olap/row.h"
 #include "olap/rowset/rowset_id_generator.h"
 #include "olap/rowset/rowset_meta_manager.h"
 #include "olap/schema_change.h"
@@ -351,72 +350,6 @@ Status PushBrokerReader::init(const Schema* schema, const TBrokerScanRange& t_sc
     }
 
     _ready = true;
-    return Status::OK();
-}
-
-Status PushBrokerReader::fill_field_row(RowCursorCell* dst, const char* src, bool src_null,
-                                        MemPool* mem_pool, FieldType type) {
-    switch (type) {
-    case OLAP_FIELD_TYPE_DECIMAL: {
-        dst->set_is_null(src_null);
-        if (src_null) {
-            break;
-        }
-        auto* decimal_value = reinterpret_cast<const DecimalV2Value*>(src);
-        auto* storage_decimal_value = reinterpret_cast<decimal12_t*>(dst->mutable_cell_ptr());
-        storage_decimal_value->integer = decimal_value->int_value();
-        storage_decimal_value->fraction = decimal_value->frac_value();
-        break;
-    }
-    case OLAP_FIELD_TYPE_DATETIME: {
-        dst->set_is_null(src_null);
-        if (src_null) {
-            break;
-        }
-
-        auto* datetime_value = reinterpret_cast<const DateTimeValue*>(src);
-        auto* storage_datetime_value = reinterpret_cast<uint64_t*>(dst->mutable_cell_ptr());
-        *storage_datetime_value = datetime_value->to_olap_datetime();
-        break;
-    }
-
-    case OLAP_FIELD_TYPE_DATE: {
-        dst->set_is_null(src_null);
-        if (src_null) {
-            break;
-        }
-
-        auto* date_value = reinterpret_cast<const DateTimeValue*>(src);
-        auto* storage_date_value = reinterpret_cast<uint24_t*>(dst->mutable_cell_ptr());
-        *storage_date_value = static_cast<int64_t>(date_value->to_olap_date());
-        break;
-    }
-    case OLAP_FIELD_TYPE_BOOL:
-    case OLAP_FIELD_TYPE_TINYINT:
-    case OLAP_FIELD_TYPE_SMALLINT:
-    case OLAP_FIELD_TYPE_INT:
-    case OLAP_FIELD_TYPE_UNSIGNED_INT:
-    case OLAP_FIELD_TYPE_BIGINT:
-    case OLAP_FIELD_TYPE_LARGEINT:
-    case OLAP_FIELD_TYPE_FLOAT:
-    case OLAP_FIELD_TYPE_DOUBLE:
-    case OLAP_FIELD_TYPE_CHAR:
-    case OLAP_FIELD_TYPE_VARCHAR:
-    case OLAP_FIELD_TYPE_STRING:
-    case OLAP_FIELD_TYPE_HLL:
-    case OLAP_FIELD_TYPE_OBJECT: {
-        dst->set_is_null(src_null);
-        if (src_null) {
-            break;
-        }
-        const auto* type_info = get_scalar_type_info(type);
-        type_info->deep_copy(dst->mutable_cell_ptr(), src, mem_pool);
-        break;
-    }
-    default:
-        return Status::Error<INVALID_SCHEMA>();
-    }
-
     return Status::OK();
 }
