+++ be/src/util/tuple_row_zorder_compare.h
@@ -19,7 +19,6 @@
 
 #include "exprs/expr.h"
 #include "exprs/expr_context.h"
-#include "olap/row.h"
 #include "olap/row_cursor.h"
 #include "olap/schema.h"
 #include "runtime/descriptors.h"
