+++ be/src/vec/olap/vgeneric_iterators.h
@@ -16,7 +16,6 @@
 // under the License.
 
 #include "olap/iterators.h"
-#include "olap/row.h"
 #include "olap/rowset/segment_v2/column_reader.h"
 
 namespace doris {
